+++ gradle-ext/main/kotlin/com/javiersc/gradle/extensions/Project.kt
@@ -10,3 +10,9 @@ public val Project.module: String
 
 public val Project.library: String
     get() = "$group:$name:$version"
+
+public val Project.isSnapshot: Boolean
+    get() = version.toString().endsWith("-SNAPSHOT", ignoreCase = true)
+
+public val Project.isNotSnapshot: Boolean
+    get() = !isSnapshot